PHP8.3以上に最適化

+++ src/cmdman/Util/cmd/repeat.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * Repeat command execution
- * @param string $cmd Command to repeat　@['require'=>true]
- * @param string $daemon PID file path
- * @param string $log Path to output the last result
- * @param string $php PHP binary path 
- * @param bool $force Forced execution
- * @param int $wt Waiting time (sec)
+ * Run a command repeatedly at a fixed interval
+ * @param string $cmd Command to execute @['require'=>true]
+ * @param int $wt Wait interval between executions (sec) @['init'=>60]
+ * @param string $daemon PID file path for daemonized execution
+ * @param string $log Output file path, or "stdout" to print directly
+ * @param string $php PHP binary path @['init'=>'php']
+ * @param bool $force Continue even if command exits with error
  */
 $php = empty($php) ? 'php' : $php;
 ob_start();
-	system($php.' -v');
+	system(escapeshellarg($php).' -v');
 $chk = ob_get_clean();
 
 if(substr($chk,0,3) !== 'PHP'){
@@ -18,69 +18,73 @@ if(substr($chk,0,3) !== 'PHP'){
 	exit;
 }
 $self = $_SERVER['PHP_SELF'];
-$command = $php.' '.$self.' '.$cmd;
+$command = escapeshellarg($php).' '.escapeshellarg($self).' '.$cmd;
 $pid = (empty($daemon) || substr($daemon,-4) == '.pid') ? $daemon : $daemon.'.pid';
 $ext_pcntl = extension_loaded('pcntl');
-$wait_status = 19;
+$wait_status = \cmdman\Util::EXIT_WAIT;
 $wait_time = empty($wt) ? 60 : $wt;
+$shutdown = false;
 
-if(!empty($log)){
-	if($log != 'stdout'){
-		\cmdman\Util::file_append($log,'');
-	}
+if(!empty($log) && $log !== 'stdout'){
+	\cmdman\Util::file_append($log,'');
 }
-$shutdown_func = function() use($pid){
-	if(file_exists($pid)){
+
+$shutdown_func = function() use($pid,&$shutdown){
+	if($shutdown){
+		return;
+	}
+	$shutdown = true;
+
+	if(!empty($pid) && file_exists($pid)){
 		unlink($pid);
 	}
 	exit;
 };
+
 if(!empty($pid)){
 	if(file_exists($pid)){
 		\cmdman\Std::println_warning('Already running');
 		exit;
-	}else{
-		\cmdman\Util::file_write($pid,(extension_loaded('posix') ? posix_getpid() : '').','.$cmd);
 	}
+	\cmdman\Util::file_write($pid,(extension_loaded('posix') ? posix_getpid() : '').','.$cmd);
+
 	if($ext_pcntl){
 		pcntl_signal(SIGINT,$shutdown_func);
 		pcntl_signal(SIGTERM,$shutdown_func);
 	}
 	register_shutdown_function($shutdown_func);
 }
+
 while(true){
 	ob_start();
 		system($command,$return_var);
 	$rtn = ob_get_clean();
 
 	if(!empty($log)){
-		if($log == 'stdout'){
+		if($log === 'stdout'){
 			print($rtn);
 		}else{
 			\cmdman\Util::file_append($log,$rtn);
 		}
-	}	
-	if(!empty($pid) && !file_exists($pid)){
-		$shutdown_func();
 	}
 	if($ext_pcntl){
 		pcntl_signal_dispatch();
+	}
+	if(!empty($pid) && !file_exists($pid)){
+		$shutdown_func();
 	}
 	if($return_var !== 0){
 		if($return_var !== $wait_status && !$force){
 			exit;
 		}
-		for($i=0,$y=0;$i<$wait_time;$i++,$y++){
+		for($i=0;$i<$wait_time;$i++){
+			sleep(1);
+
 			if($ext_pcntl){
 				pcntl_signal_dispatch();
 			}
-			sleep(1);
-			
-			if($y == 60){
-				if(!empty($pid) && !file_exists($pid)){
-					$shutdown_func();
-				}
-				$y = 0;
+			if(!empty($pid) && $i % 60 === 59 && !file_exists($pid)){
+				$shutdown_func();
 			}
 		}
 	}
